appbar is functional
