+++ src/Cart.css
@@ -10,3 +10,7 @@
 .cart-items {
   padding-top: 5rem;
 }
+
+.cartt {
+  font-size: 0;
+}
